Remove dead code, reorganize files
