--- src/components/game/dice/Dice.css
@@ -1,7 +1,0 @@
-.dice-container {
-  position: absolute;
-  top: 0px;
-  left: 0px;
-  padding: 10px;
-  z-index: 2;
-}
