--- src/components/game/dice/DiceButton.css
@@ -1,17 +1,0 @@
-.dice-button-background {
-  background-color: #cc3716;
-  border-radius: 50%;
-  cursor: pointer;
-  width: 50px;
-  height: 50px;
-  padding: 3px;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-}
-
-.dice-button-icon {
-  cursor: pointer;
-  width: 60%;
-  height: 60%;
-}
